Move is_cloud_url to utils
