+++ crates/polars-io/src/lib.rs
@@ -13,33 +13,29 @@ pub mod export;
 pub mod ipc;
 #[cfg(feature = "json")]
 pub mod json;
+pub mod mmap;
 #[cfg(feature = "json")]
 pub mod ndjson;
-#[cfg(feature = "cloud")]
-pub use cloud::glob as async_glob;
-
-pub mod mmap;
 mod options;
 #[cfg(feature = "parquet")]
 pub mod parquet;
+#[cfg(feature = "partition")]
+pub mod partition;
+#[cfg(feature = "async")]
+pub mod pl_async;
 pub mod predicates;
 pub mod prelude;
 #[cfg(all(test, feature = "csv"))]
 mod tests;
 pub mod utils;
-use once_cell::sync::Lazy;
-use regex::Regex;
-
-#[cfg(feature = "partition")]
-pub mod partition;
-#[cfg(feature = "async")]
-pub mod pl_async;
 
 use std::io::{Read, Write};
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 #[allow(unused)] // remove when updating to rust nightly >= 1.61
 use arrow::array::new_empty_array;
+#[cfg(feature = "cloud")]
+pub use cloud::glob as async_glob;
 pub use options::*;
 use polars_core::frame::ArrowChunk;
 use polars_core::prelude::*;
@@ -150,15 +146,4 @@ pub(crate) fn finish_reader<R: ArrowReader>(
     };
 
     Ok(if rechunk { df.agg_chunks() } else { df })
-}
-
-static CLOUD_URL: Lazy<Regex> =
-    Lazy::new(|| Regex::new(r"^(s3a?|gs|gcs|file|abfss?|azure|az|adl|https?)://").unwrap());
-
-/// Check if the path is a cloud url.
-pub fn is_cloud_url<P: AsRef<Path>>(p: P) -> bool {
-    match p.as_ref().as_os_str().to_str() {
-        Some(s) => CLOUD_URL.is_match(s),
-        _ => false,
-    }
 }
