+++ crates/polars-pipe/src/executors/sources/parquet.rs
@@ -16,8 +16,8 @@ use polars_io::predicates::PhysicalIoExpr;
 use polars_io::prelude::materialize_projection;
 #[cfg(feature = "async")]
 use polars_io::prelude::ParquetAsyncReader;
-use polars_io::utils::check_projected_arrow_schema;
-use polars_io::{is_cloud_url, SerReader};
+use polars_io::utils::{check_projected_arrow_schema, is_cloud_url};
+use polars_io::SerReader;
 use polars_plan::logical_plan::FileInfo;
 use polars_plan::prelude::FileScanOptions;
 use polars_utils::iter::EnumerateIdxTrait;
